Compile and commit all of the LLVM tools, including opt, llc, and lld
This commit compiles all of the LLVM tools with _malloc, _free, the FS
object, stringToUTF8, and UTF8ToString exposed to JavaScript. Note that
wasm-ld.js was just a symlink to lld.js, and hence it was not included
in this commit. Also, llvm-lit and llvm-locstats (whatever they are)
were not included, because they were compiled to native for whatever
reason, and they conveniently lack file extensions!